src/requirements.txt
# Django Starter Dependencies
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
celery>=5.3
redis>=5.0
psycopg2-binary>=2.9
dj-database-url>=2.0
python-dotenv>=1.0
gunicorn>=21.0
whitenoise>=6.5
Pillow>=10.0
django-filter>=23.0
drf-yasg>=1.21
sentry-sdk>=1.40
